Some checks

+++ compute/make-proxmox-iso.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 set -eu
 cd proxmox-auto-install-assistant-docker
-mkdir iso
-wget -O iso/proxmox-ve_9.2-1.iso https://enterprise.proxmox.com/iso/proxmox-ve_9.2-1.iso
-docker build -t proxmox-auto-installer .
+
+if [[ ! -d iso]]; then
+    mkdir iso
+fi
+[ -f iso/proxmox-ve_9.2-1.iso ] || wget -O iso/proxmox-ve_9.2-1.iso https://enterprise.proxmox.com/iso/proxmox-ve_9.2-1.isodocker build -t proxmox-auto-installer .
 chmod +x entrypoint.sh
 docker run --rm \
   -v $PWD/iso:/iso:ro \
